% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neg_bin.R
\name{test_call_to_tfd}
\alias{test_call_to_tfd}
\title{Add together two numbers}
\usage{
test_call_to_tfd()
}
\value{
The sum of `x` and `y`
}
\description{
Add together two numbers
}
\examples{

test_call_to_tfd()
}
