% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neg_bin.R
\name{glm_negbin}
\alias{glm_negbin}
\title{Add together two numbers}
\usage{
glm_negbin(thedata = NULL)
}
\arguments{
\item{thedata}{data.frame of the data to fit to the model}
}
\value{
currently nothing
}
\description{
A short description...
}
\examples{
\dontrun{
library(rstanarm)
data(roaches)
roaches$roach1<-roaches$roach1/100;# manual
glm_negbin(thedata=roaches)
}
}
